feat: encapsulate searchable events section component
