+++ src/components/EventsSearch/index.js
@@ -1,0 +1,3 @@
+import EventsSearch from './EventsSearch';
+
+export default EventsSearch;
